# api-docs.json
{"openapi":"3.1.0","info":{"title":"My API","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/{resource}":{"get":{"tags":["authorization-controller"],"operationId":"isAuthorizedToGet","parameters":[{"name":"resource","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User is authorized to get resource"},"403":{"description":"User is not authorized to get resource"},"400":{"description":"Invalid resource"}},"security":[{"bearerAuth":[]}]},"put":{"tags":["authorization-controller"],"operationId":"isAuthorizedToPut","parameters":[{"name":"resource","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User is authorized to put resource"},"403":{"description":"User is not authorized to put resource"},"400":{"description":"Invalid resource"}},"security":[{"bearerAuth":[]}]},"post":{"tags":["authorization-controller"],"operationId":"isAuthorizedToCreate","parameters":[{"name":"resource","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User is authorized to create resource"},"403":{"description":"User is not authorized to create resource"},"400":{"description":"Invalid resource"}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["authorization-controller"],"operationId":"isAuthorizedToDelete","parameters":[{"name":"resource","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User is authorized to delete resource"},"403":{"description":"User is not authorized to delete resource"},"400":{"description":"Invalid resource"}},"security":[{"bearerAuth":[]}]},"patch":{"tags":["authorization-controller"],"operationId":"isAuthorizedToPatch","parameters":[{"name":"resource","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User is authorized to patch resource"},"403":{"description":"User is not authorized to patch resource"},"400":{"description":"Invalid resource"}},"security":[{"bearerAuth":[]}]}},"/users/{id}":{"get":{"tags":["users-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User retrieved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetUserResponse"}}}},"400":{"description":"Invalid request or user not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["users-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"204":{"description":"User updated"},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["users-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User disabled successfully","content":{"*/*":{"schema":{"type":"string","default":"User disabled successfully"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"Failed to disable user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"patch":{"tags":["users-controller"],"operationId":"updateUserPassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserPasswordRequest"}}},"required":true},"responses":{"204":{"description":"User password updated"},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]}},"/roles/{role-name}":{"get":{"tags":["roles-controller"],"operationId":"getRole","parameters":[{"name":"role-name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Role retrieved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleRepresentation"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"Role not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["roles-controller"],"operationId":"updateRole","parameters":[{"name":"role-name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleRequest"}}},"required":true},"responses":{"204":{"description":"Role updated","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"Role not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["roles-controller"],"operationId":"deleteRole","parameters":[{"name":"role-name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Role deleted","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"Role not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"patch":{"tags":["roles-controller"],"operationId":"patchRole","parameters":[{"name":"role-name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchRoleRequest"}}},"required":true},"responses":{"204":{"description":"Role updated","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"Role not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]}},"/users":{"get":{"tags":["users-controller"],"operationId":"getUsers","parameters":[{"name":"username","in":"query","required":false,"schema":{"type":"string"}},{"name":"firstName","in":"query","required":false,"schema":{"type":"string"}},{"name":"lastName","in":"query","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"enabled","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"Users retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetUserResponse"}}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["users-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"201":{"description":"User created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateUserResponse"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"409":{"description":"User already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]}},"/users/{id}/role-mappings":{"get":{"tags":["users-controller"],"operationId":"getUserRoleMappings","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User role mappings retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleRepresentation"}}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["users-controller"],"operationId":"createUserRoleMappings","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"201":{"description":"Role mappings created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["users-controller"],"operationId":"deleteUserRoleMappings","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}},"required":true},"responses":{"204":{"description":"Role mappings deleted","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]}},"/roles":{"get":{"tags":["roles-controller"],"operationId":"getRoles","responses":{"200":{"description":"Roles retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleRepresentation"}}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["roles-controller"],"operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleRequest"}}},"required":true},"responses":{"201":{"description":"Role created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleRepresentation"}}}},"400":{"description":"Invalid request or email","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"403":{"description":"Access token lacks required admin scopes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"409":{"description":"Role already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]}},"/login":{"post":{"tags":["authentication-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccessToken"}}}},"401":{"description":"Invalid email or password","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}}}},"/login/refresh":{"post":{"tags":["authentication-controller"],"operationId":"loginWithRefreshToken","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}],"responses":{"200":{"description":"Login with refresh token successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccessToken"}}}},"400":{"description":"Could not parse refresh token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"401":{"description":"Invalid refresh token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}},"500":{"description":"An unexpected error occurred","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}}}},"/users/current":{"get":{"tags":["users-controller"],"operationId":"getCurrentUser","responses":{"200":{"description":"User retrieved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/KeycloakUserInfo"}}}},"401":{"description":"Invalid access token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OAuthError"}}}}},"security":[{"bearerAuth":[]}]}},"/health":{"get":{"tags":["health-check-controller"],"operationId":"health","responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UpdateUserRequest":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"}}},"OAuthError":{"type":"object","properties":{"errorDescription":{"type":"string"},"source":{"type":"string"}}},"CreateRoleRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["description","name"]},"CreateUserRequest":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"}}},"CreateUserResponse":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"id":{"type":"string"}}},"Composites":{"type":"object","properties":{"realm":{"type":"array","items":{"type":"string"},"uniqueItems":true},"client":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}},"application":{"type":"object","additionalProperties":{"type":"array","deprecated":true,"items":{"type":"string","deprecated":true}},"deprecated":true}}},"RoleRepresentation":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"scopeParamRequired":{"type":"boolean","deprecated":true},"composite":{"type":"boolean"},"composites":{"$ref":"#/components/schemas/Composites"},"clientRole":{"type":"boolean"},"containerId":{"type":"string"},"attributes":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AccessToken":{"type":"object","properties":{"access_token":{"type":"string"},"expires_in":{"type":"integer","format":"int32"},"refresh_expires_in":{"type":"integer","format":"int32"},"refresh_token":{"type":"string"},"token_type":{"type":"string"},"scope":{"type":"string"}}},"RefreshTokenRequest":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"UpdateUserPasswordRequest":{"type":"object","properties":{"newPassword":{"type":"string"}}},"PatchRoleRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"GetUserResponse":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"enabled":{"type":"boolean"},"id":{"type":"string"}}},"KeycloakUserInfo":{"type":"object","properties":{"sub":{"type":"string"},"name":{"type":"string"},"scope":{"type":"string"},"given_name":{"type":"string"},"family_name":{"type":"string"},"preferred_username":{"type":"string"},"email":{"type":"string"},"picture":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}